requirements.txt
pycryptodome
gotify-handler
parsedatetime
python-daemon
requests
schedule
yagmail
ebaysdk
python-dotenv
